refactor: 募集関係のキャンバスの関数の移動
