+++ app/canvas/recruit/big_run_canvas.js
@@ -1,13 +1,13 @@
+const Canvas = require('canvas');
+const { createRoundRect, drawArcImage, fillTextWithStroke } = require('../../common/canvas_components');
+const log4js = require('log4js');
+const { getBigRunData } = require('../../common/apis/splatoon3_ink');
+const { dateformat, formatDatetime } = require('../../common/convert_datetime');
+
 module.exports = {
     recruitBigRunCanvas: recruitBigRunCanvas,
     ruleBigRunCanvas: ruleBigRunCanvas,
 };
-
-const Canvas = require('canvas');
-const { createRoundRect, drawArcImage, fillTextWithStroke } = require('../../../common/canvas_components');
-const log4js = require('log4js');
-const { getBigRunData } = require('../../../common/apis/splatoon3_ink');
-const { dateformat, formatDatetime } = require('../../../common/convert_datetime');
 
 log4js.configure(process.env.LOG4JS_CONFIG_PATH);
 const logger = log4js.getLogger('recruit');
