+++ app/cmd/splat3/recruit/salmon_recruit.js
@@ -1,29 +1,15 @@
-const Canvas = require('canvas');
-const path = require('path');
 const { searchMessageById } = require('../../../manager/messageManager');
 const { searchMemberById } = require('../../../manager/memberManager');
 const { isNotEmpty, sleep } = require('../../../common');
-const { createRoundRect, drawArcImage, fillTextWithStroke } = require('../../../common/canvas_components');
 const { recruitActionRow, recruitDeleteButton, unlockChannelButton } = require('../../../buttons/create_recruit_buttons');
 const { AttachmentBuilder, PermissionsBitField } = require('discord.js');
 const log4js = require('log4js');
-const { fetchSchedule, getSalmonData, checkBigRun } = require('../../../common/apis/splatoon3_ink');
-const { dateformat, formatDatetime } = require('../../../common/convert_datetime');
-const { recruitBigRunCanvas, ruleBigRunCanvas } = require('./big_run_recruit');
+const { fetchSchedule, checkBigRun } = require('../../../common/apis/splatoon3_ink');
+const { recruitBigRunCanvas, ruleBigRunCanvas } = require('../../../canvas/recruit/big_run_canvas');
+const { recruitSalmonCanvas, ruleSalmonCanvas } = require('../../../canvas/recruit/salmon_canvas');
 
 log4js.configure(process.env.LOG4JS_CONFIG_PATH);
 const logger = log4js.getLogger('recruit');
-
-Canvas.registerFont(path.resolve('./fonts/Splatfont.ttf'), {
-    family: 'Splatfont',
-});
-Canvas.registerFont(path.resolve('./fonts/GenShinGothic-P-Medium.ttf'), {
-    family: 'Genshin',
-});
-Canvas.registerFont(path.resolve('./fonts/GenShinGothic-P-Bold.ttf'), {
-    family: 'Genshin-Bold',
-});
-Canvas.registerFont(path.resolve('./fonts/SEGUISYM.TTF'), { family: 'SEGUI' });
 
 module.exports = {
     salmonRecruit: salmonRecruit,
@@ -132,14 +118,14 @@ async function sendSalmonRun(interaction, txt, recruit_num, condition, count, ho
     if (checkBigRun(data.schedule, 0)) {
         recruitBuffer = await recruitBigRunCanvas(recruit_num, count, host_member, user1, user2, condition, channel_name);
     } else {
-        recruitBuffer = await recruitCanvas(recruit_num, count, host_member, user1, user2, condition, channel_name);
+        recruitBuffer = await recruitSalmonCanvas(recruit_num, count, host_member, user1, user2, condition, channel_name);
     }
 
     let ruleBuffer;
     if (checkBigRun(data.schedule, 0)) {
         ruleBuffer = await ruleBigRunCanvas(data);
     } else {
-        ruleBuffer = await ruleCanvas(data);
+        ruleBuffer = await ruleSalmonCanvas(data);
     }
 
     const recruit = new AttachmentBuilder(recruitBuffer, 'ikabu_recruit.png');
@@ -216,187 +202,4 @@ async function sendSalmonRun(interaction, txt, recruit_num, condition, count, ho
     } catch (error) {
         logger.error(error);
     }
-}
-
-/*
- * 募集用のキャンバス(1枚目)を作成する
- */
-async function recruitCanvas(recruit_num, count, host_member, user1, user2, condition, channel_name) {
-    blank_avatar_url = 'https://raw.githubusercontent.com/shngmsw/ikabu/main/images/recruit/blank_avatar.png'; // blankのアバター画像URL
-
-    const recruitCanvas = Canvas.createCanvas(720, 550);
-    const recruit_ctx = recruitCanvas.getContext('2d');
-
-    // 下地
-    createRoundRect(recruit_ctx, 1, 1, 718, 548, 30);
-    recruit_ctx.fillStyle = '#2F3136';
-    recruit_ctx.fill();
-    recruit_ctx.strokeStyle = '#FFFFFF';
-    recruit_ctx.lineWidth = 4;
-    recruit_ctx.stroke();
-
-    let salmon_icon = await Canvas.loadImage('https://raw.githubusercontent.com/shngmsw/ikabu/main/images/recruit/salmon_black_icon.png');
-    recruit_ctx.drawImage(salmon_icon, 22, 32, 82, 60);
-
-    fillTextWithStroke(recruit_ctx, 'SALMON', '51px Splatfont', '#000000', '#FF9900', 3, 115, 80);
-    fillTextWithStroke(recruit_ctx, 'RUN', '51px Splatfont', '#000000', '#00FF00DA', 3, 350, 80);
-
-    // 募集主の画像
-    let host_img = await Canvas.loadImage(host_member.displayAvatarURL({ extension: 'png' }));
-    recruit_ctx.save();
-    drawArcImage(recruit_ctx, host_img, 40, 120, 50);
-    recruit_ctx.strokeStyle = '#1e1f23';
-    recruit_ctx.lineWidth = 9;
-    recruit_ctx.stroke();
-    recruit_ctx.restore();
-
-    let user1_url = blank_avatar_url;
-    let user2_url = blank_avatar_url;
-    let user3_url = blank_avatar_url;
-
-    // 参加者指定があれば、画像を拾ってくる
-    if (user1 != null && user2 != null) {
-        user1_url = user1.displayAvatarURL({ extension: 'png' });
-        user2_url = user2.displayAvatarURL({ extension: 'png' });
-    } else if (user1 != null && user2 == null) {
-        user1_url = user1.displayAvatarURL({ extension: 'png' });
-    } else if (user1 == null && user2 != null) {
-        user1_url = user2.displayAvatarURL({ extension: 'png' });
-    }
-
-    let user1_img = await Canvas.loadImage(user1_url);
-    recruit_ctx.save();
-    drawArcImage(recruit_ctx, user1_img, 158, 120, 50);
-    recruit_ctx.strokeStyle = '#1e1f23';
-    recruit_ctx.lineWidth = 9;
-    recruit_ctx.stroke();
-    recruit_ctx.restore();
-
-    if (count >= 3) {
-        let user2_img = await Canvas.loadImage(user2_url);
-        recruit_ctx.save();
-        drawArcImage(recruit_ctx, user2_img, 276, 120, 50);
-        recruit_ctx.strokeStyle = '#1e1f23';
-        recruit_ctx.lineWidth = 9;
-        recruit_ctx.stroke();
-        recruit_ctx.restore();
-    }
-
-    if (count == 4) {
-        let user3_img = await Canvas.loadImage(user3_url);
-        recruit_ctx.save();
-        drawArcImage(recruit_ctx, user3_img, 394, 120, 50);
-        recruit_ctx.strokeStyle = '#1e1f23';
-        recruit_ctx.lineWidth = 9;
-        recruit_ctx.stroke();
-        recruit_ctx.restore();
-    }
-
-    let host_icon = await Canvas.loadImage('https://raw.githubusercontent.com/shngmsw/ikabu/main/images/recruit/squid.png');
-    recruit_ctx.drawImage(host_icon, 0, 0, host_icon.width, host_icon.height, 90, 172, 75, 75);
-
-    fillTextWithStroke(recruit_ctx, '募集人数', '39px "Splatfont"', '#FFFFFF', '#2D3130', 1, 525, 155);
-
-    fillTextWithStroke(recruit_ctx, '@' + recruit_num, '42px "Splatfont"', '#FFFFFF', '#2D3130', 1, 580, 218);
-
-    fillTextWithStroke(recruit_ctx, '参加条件', '43px "Splatfont"', '#FFFFFF', '#2D3130', 1, 35, 290);
-
-    recruit_ctx.font = '30px "Genshin", "SEGUI"';
-    const width = 600;
-    const size = 40;
-    const column_num = 4;
-    let column = [''];
-    let line = 0;
-    condition = condition.replace('{br}', '\n', 'gm');
-
-    // 幅に合わせて自動改行
-    for (var i = 0; i < condition.length; i++) {
-        var char = condition.charAt(i);
-
-        if (char == '\n') {
-            line++;
-            column[line] = '';
-        } else if (recruit_ctx.measureText(column[line] + char).width > width) {
-            line++;
-            column[line] = char;
-        } else {
-            column[line] += char;
-        }
-    }
-
-    if (column.length > column_num) {
-        column[column_num - 1] += '…';
-    }
-
-    for (var j = 0; j < column.length; j++) {
-        if (j < column_num) {
-            recruit_ctx.fillText(column[j], 65, 345 + size * j);
-        }
-    }
-
-    fillTextWithStroke(recruit_ctx, channel_name, '37px "Splatfont"', '#FFFFFF', '#2D3130', 1, 30, 520);
-
-    const recruit = recruitCanvas.toBuffer();
-    return recruit;
-}
-
-/*
- * ルール情報のキャンバス(2枚目)を作成する
- */
-async function ruleCanvas(data) {
-    const salmon_data = await getSalmonData(data, 0);
-
-    const datetime =
-        formatDatetime(salmon_data.startTime, dateformat.mdwhm) + ' - ' + formatDatetime(salmon_data.endTime, dateformat.mdwhm);
-
-    const ruleCanvas = Canvas.createCanvas(720, 550);
-    const rule_ctx = ruleCanvas.getContext('2d');
-
-    createRoundRect(rule_ctx, 1, 1, 718, 548, 30);
-    rule_ctx.fillStyle = '#2F3136';
-    rule_ctx.fill();
-    rule_ctx.strokeStyle = '#FFFFFF';
-    rule_ctx.lineWidth = 4;
-    rule_ctx.stroke();
-
-    fillTextWithStroke(rule_ctx, '日時', '32px Splatfont', '#FFFFFF', '#2D3130', 1, 35, 80);
-
-    date_width = rule_ctx.measureText(datetime).width;
-    fillTextWithStroke(rule_ctx, datetime, '37px Splatfont', '#FFFFFF', '#2D3130', 1, (650 - date_width) / 2, 145);
-
-    fillTextWithStroke(rule_ctx, '武器', '32px Splatfont', '#FFFFFF', '#2D3130', 1, 35, 245);
-
-    let weapon1_img = await Canvas.loadImage(salmon_data.weapon1);
-    rule_ctx.drawImage(weapon1_img, 50, 280, 110, 110);
-
-    let weapon2_img = await Canvas.loadImage(salmon_data.weapon2);
-    rule_ctx.drawImage(weapon2_img, 190, 280, 110, 110);
-
-    let weapon3_img = await Canvas.loadImage(salmon_data.weapon3);
-    rule_ctx.drawImage(weapon3_img, 50, 410, 110, 110);
-
-    let weapon4_img = await Canvas.loadImage(salmon_data.weapon4);
-    rule_ctx.drawImage(weapon4_img, 190, 410, 110, 110);
-
-    fillTextWithStroke(rule_ctx, 'ステージ', '33px Splatfont', '#FFFFFF', '#2D3130', 1, 350, 245);
-
-    stage_width = rule_ctx.measureText(salmon_data.stage).width;
-    fillTextWithStroke(rule_ctx, salmon_data.stage, '38px Splatfont', '#FFFFFF', '#2D3130', 1, 150 + (700 - stage_width) / 2, 300);
-
-    let stage_img = await Canvas.loadImage(salmon_data.stageImage);
-    rule_ctx.save();
-    rule_ctx.beginPath();
-    createRoundRect(rule_ctx, 370, 340, 308, 176, 10);
-    rule_ctx.clip();
-    rule_ctx.drawImage(stage_img, 370, 340, 308, 176);
-    rule_ctx.strokeStyle = '#FFFFFF';
-    rule_ctx.lineWidth = 6.0;
-    rule_ctx.stroke();
-    rule_ctx.restore();
-
-    createRoundRect(rule_ctx, 1, 1, 718, 548, 30);
-    rule_ctx.clip();
-
-    const rule = ruleCanvas.toBuffer();
-    return rule;
 }
